+++ src/merging/merging_nnls.jl
@@ -74,21 +74,21 @@ mutable struct NNLSMerge
     scaley::Float64
     scalez::Float64
 
-    n_total_conserved::Int32
-    n_moments_vel::Int32
+    n_total_conserved::Int64
+    n_moments_vel::Int64
     rhs_vector::Vector{Float64}
-    mim::Vector{Vector{Int32}}  # mult-index moments
-    tot_order::Vector{Int32}
-    n_moments_pos::Int32
-    mim_pos::Vector{Vector{Int32}}  # mult-index moments
-    tot_order_pos::Vector{Int32}
-    pos_i_x::Int32
-    pos_i_y::Int32
-    pos_i_z::Int32
+    mim::Vector{Vector{Int64}}  # mult-index moments
+    tot_order::Vector{Int64}
+    n_moments_pos::Int64
+    mim_pos::Vector{Vector{Int64}}  # mult-index moments
+    tot_order_pos::Vector{Int64}
+    pos_i_x::Int64
+    pos_i_y::Int64
+    pos_i_z::Int64
 
     lhs_matrices::Vector{Matrix{Float64}}
-    lhs_matrix_ncols_start::Int32
-    lhs_matrix_ncols_end::Int32
+    lhs_matrix_ncols_start::Int64
+    lhs_matrix_ncols_end::Int64
     column_norms::Vector{Vector{Float64}}
     vel_pos_matrices::Vector{Matrix{Float64}}
